requirements.txt
opencv-python
numpy
ultralytics
transformers
pillow
python-dotenv
requests
torch
torchvision
